requirements.txt
Flask==2.3.2
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.27.0
python-dotenv>=1.0.0
